[ADD/IMP]payment_export_text: made compatible with base export engine ref:issues/941
