+++ base_payment_export/models/__init__.py
@@ -1,3 +1,2 @@
 # -*- coding: utf-8 -*-
-# from . import payment_export
 from . import payment_export_config
